feat: fetchLyrics
